% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\docType{data}
\name{my_colors}
\alias{my_colors}
\title{color list}
\format{An object of class \code{list} of length 6.}
\usage{
my_colors
}
\description{
list of custom colors for palettes
}
\keyword{datasets}
