% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_palette.R
\name{my_palette}
\alias{my_palette}
\title{palette function}
\usage{
my_palette(primary = "orange", other = "gold", direction = 1)
}
\arguments{
\item{primary}{initial color - must be in the selection}

\item{other}{if selection only 2 colors, the secondary color}

\item{direction}{whether you want ot reverse the direction of the colors}
}
\description{
baed on functions published by Garrick Aden-Buie
}
\examples{
ggplot() + my_palette(direction = -1)(2)
}
\keyword{palette}
