% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_matt}
\alias{scale_fill_matt}
\title{scale_fill_matt}
\usage{
scale_fill_matt(primary = "orange", other = "plum", direction = 1,
  ...)
}
\description{
This function applies my_palette
}
\examples{
ggplot() + scale_fill_matt()
}
